Factor out locale handling into a separate utility module.
